Stop cpuid from sign-extending every returned register
__cpuidex writes into an int[4], and the results were widened straight into
the 64-bit cells, so any register with bit 31 set came back as
0xFFFFFFFF________. CPUID.0x80000000:EAX is 0x8000000x and CPUID.1:ECX[31]
is the hypervisor-present bit, so this was not theoretical: bit tests still
worked, but shifts, comparisons and equality against a 32-bit constant did
not. Every other native converts through an unsigned type.

+++ PawnIO/src/natives_impl_windows.cpp
@@ -463,7 +463,7 @@ void cpuid(cell leaf, cell subleaf, std::array<cell, 4>& out) {
   int out32[4];
   __cpuidex(out32, (int)leaf, (int)subleaf);
   for (size_t i = 0; i < 4; ++i)
-    out[i] = out32[i];
+    out[i] = (uint32_t)out32[i];
 }
 
 extern "C" cell _crdr(cell id, cell v);
